chore(cosmwasm): modular tests
